+++ packages/frontend/src/components/Settings/Advanced.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect, useState, useCallback } from 'react'
 
 import { SettingsStoreState } from '../../stores/settings'
 import Encryption from './Encryption'
@@ -12,6 +12,8 @@ import EditAccountAndPasswordDialog from '../dialogs/EditAccountAndPasswordDialo
 import useDialog from '../../hooks/dialog/useDialog'
 import SettingsButton from './SettingsButton'
 import { runtime } from '@deltachat-desktop/runtime-interface'
+import DesktopSettingsSwitch from './DesktopSettingsSwitch'
+import { AutostartState } from '@deltachat-desktop/shared/shared-types'
 
 type Props = {
   settingsStore: SettingsStoreState
@@ -26,6 +28,10 @@ export default function Advanced({ settingsStore }: Props) {
       <SettingsButton onClick={() => runtime.openLogFile()}>
         {tx('pref_view_log')}
       </SettingsButton>
+      {/*
+        don't show it on electron yet, as the message "not available on this runtime/platform"
+        would confuse users as long as tauri is not the default */}
+      {runtime.getRuntimeInfo().target === 'tauri' && <SettingsAutoStart />}
       <Communication settingsStore={settingsStore} />
       <SettingsSeparator />
 
@@ -50,6 +56,44 @@ export default function Advanced({ settingsStore }: Props) {
       {settingsStore.settings.is_chatmail == '0' && (
         <ImapFolderHandling settingsStore={settingsStore} />
       )}
+    </>
+  )
+}
+
+function SettingsAutoStart() {
+  const tx = useTranslationFunction()
+
+  const [autostartState, setAutostartState] = useState<AutostartState | null>(
+    null
+  )
+  const update = useCallback(() => {
+    runtime.getAutostartState().then(setAutostartState)
+  }, [])
+
+  useEffect(() => {
+    update()
+  }, [update])
+
+  return (
+    <>
+      <DesktopSettingsSwitch
+        // force react to rerender element, so that the switch does not animate
+        key={String(!autostartState)}
+        settingsKey='autostart'
+        label={tx('pref_autostart')}
+        description={
+          autostartState
+            ? autostartState.isSupported
+              ? autostartState.isRegistered
+                ? tx('pref_autostart_registered')
+                : tx('pref_autostart_not_registered')
+              : tx('pref_autostart_not_supported')
+            : undefined // don't show description while it is loading
+        }
+        disabled={!autostartState?.isSupported}
+        disabledValue={false}
+        callback={update}
+      />
     </>
   )
 }