+++ packages/target-tauri/runtime-tauri/runtime.ts
@@ -10,7 +10,8 @@ import type { Store } from '@tauri-apps/plugin-store'
 import { openPath, openUrl } from '@tauri-apps/plugin-opener'
 import { writeText, readText } from '@tauri-apps/plugin-clipboard-manager'
 
-import {
+import type {
+  AutostartState,
   DcNotification,
   DcOpenWebxdcParameters,
   DesktopSettingsType,
@@ -21,7 +22,10 @@ import {
 } from '@deltachat-desktop/shared/shared-types.js'
 import '@deltachat-desktop/shared/global.d.ts'
 
-import { Runtime, RuntimeAppPath } from '@deltachat-desktop/runtime-interface'
+import type {
+  Runtime,
+  RuntimeAppPath,
+} from '@deltachat-desktop/runtime-interface'
 import { BaseDeltaChat, yerpc } from '@deltachat/jsonrpc-client'
 import type { LocaleData } from '@deltachat-desktop/shared/localize.js'
 import type {
@@ -120,6 +124,7 @@ class TauriRuntime implements Runtime {
       HTMLEmailAlwaysLoadRemoteContent: false,
       contentProtectionEnabled: false,
       locale: null, // if this is null, the system chooses the system language that electron reports
+      autostart: true,
     } satisfies Partial<DesktopSettingsType>
 
     const frontendOnly = {
@@ -494,6 +499,9 @@ class TauriRuntime implements Runtime {
   }
   isDroppedFileFromOutside(_file: File): boolean {
     throw new Error('Method not implemented.51')
+  }
+  getAutostartState(): Promise<AutostartState> {
+    return invoke('get_autostart_state')
   }
   onChooseLanguage: ((locale: string) => Promise<void>) | undefined
   onThemeUpdate: (() => void) | undefined