+++ packages/target-tauri/src-tauri/src/lib.rs
@@ -7,7 +7,9 @@ use state::{
 };
 use tauri::Manager;
 use util::csp::add_custom_schemes_to_csp_for_window_and_android;
+
 mod app_path;
+mod autostart;
 mod blobs;
 mod clipboard;
 mod file_dialogs;
@@ -101,6 +103,7 @@ pub fn run() {
             ui_frontend_ready,
             get_current_logfile,
             copy_image_to_clipboard,
+            autostart::get_autostart_state,
             app_path::get_app_path,
             clipboard::get_clipboard_image_as_data_uri,
             file_dialogs::download_file,
@@ -174,6 +177,18 @@ pub fn run() {
                 let _ = tauri_plugin_log::attach_logger(max_level, logger);
             }
             app.handle().plugin(tauri_plugin_log)?;
+
+            #[cfg(desktop)]
+            {
+                use tauri_plugin_autostart::MacosLauncher;
+
+                app.handle().plugin(tauri_plugin_autostart::init(
+                    MacosLauncher::LaunchAgent,
+                    Some(vec!["--minimized", "--autostart"]),
+                    // TODO: note that cli arguments are not implemented yet
+                    // TODO: `--autostart` should show a different message why the tray option is disabled
+                ))?;
+            }
 
             app.manage(tauri::async_runtime::block_on(AppState::try_new(
                 app,